movie_admin/requirements/dev.txt
-r base.txt
django-debug-toolbar==2.2
pylint==2.8.2